Use client_secret_post authentication method
Changes so that the azure app client secret is used when authenticating
users. In order to enable this, the AZURE_CLIENT_SECRET needs to be
added as an environment variable. Also need to update the authentication
urls within Azure portal, so that the authenticate url is registered as
a web redirect URI.