Feature/embed quicksight (#233)
* Code to exchange entra token for aws token to test identity propogation

* Added commands to get local debugging working against apc dev. Added JWT as dependency

* Added quicksight embedding

* Add initial database access app

* Move quicksight template, use namespace url

* Use management account profile when running locally

* Initial aws service for Quicksight

* Add code to ensure single boto3 session created

* Add glueservice and more refactoring

* Refactor base AWS service

* Fix tests and linting

* Remove static analysis feature from devcontainer

Static analysis includes checkov, which installed a version of botocore
that conflicted with the version pinned in the requirements. It also
installed packages as the root user, which meant when trying to upgrade
botocore, a permission error was raised as the vscode user that runs the
dev container does not have permission to remove packages installed as
root. We dont need to run checkov locally so removing, which resolves
the issue.

* Add instructions to retreive .env file

* Attempt to appease superlinter

* Update the home and login failure pages

* Tidy up the frontend

* Prompt user to login when authenticating

* Update default behaviour when retreiving STS creds

We only want to default to using STS to retrieve AWS credentials when
not using a service role that already has permission to call services.
These changes attempt to make that more explicit. For now, we should
only default to using STS when running locally.

---------

Co-authored-by: jamesstottmoj <james.stott@digital.justice.gov.uk>