Feature/manage databases (#250)
* Add LakeFormation service

Methods to grant access to databases and tables.
Also update glue service to add method to return target table data.

* Add models and views to create database access

* Rename GrantTableAccessView

* Raise 404 if table/database not found

* Limit grantable access based on the users access

A user should only be able to grant access to another user that they
themselves have. If the user is a superuser, they can grant any access.

* Bad idea about using an UpdateView

* Revert "Bad idea about using an UpdateView"

This reverts commit 333b0a6c95600c711cf3cd61f4d5ecb271ab30a0.

* Add view to manage existing table access

* Add django-debug-toolbar for local running

* Add view to revoke access to a table

* Grant/revoke LF access when objects are saved/deleted

* Fix HTML bugs

* Add debug toolbar to requirements so can be used in dev environment

* Render form using includes

* Create custom template for select field

* Remove form includes

* Add custom template for checkbox field

* Update select template, add checkbox template

* Try to fix HMTL lint error

* Reformat file

* Remove validation error added for debugging

* Update granting access call

Dont save the model if the API call fails. Also
use the values from the selected accesslevel model
when making API call.

* Create hybrid opt-in when granting access

Move logic away from model save methods, and on
to methods on the model that are called by the
access forms.

* Handle errors when granting permissions

Use DB transactions to keep the database state in
sync with lake formation permissions.
Add validation to check that selected grantable
are valid based on standard permissions.

* Add display_name to AccessLevel model

* Store permissions and grantable perms separately

Refactors code to remove need for duplicate access levels with
grantable flag.

* Rename AccessLevel to Permission

Remove the grantable field as no longer necessary

* Give users DESCRIBE grantable access to the DB

When a user is granted access to a table, grant the user DESCRIBE access
to the database with grantable permission. This is because having access
to describe the database only allows you to see the name of the DB.

* Squash migrations

* Add default PERMISSION objects

* Hide manage access buttons when neccessary

Only show the actions column to manage/revoke table access if the user
is either a superuser, or has grantable permissions on the table.

* Add admin management for users

* Disable some extra linters

* Update helptext for grantable permissions