Update user usernames
When creating users, use the UPN, preferred username or email address.
This will only occur on initial login. For existing users, a data
migration has been addeded to update the usernames to their email
address, with the domain lowercase.
This resolves a bug where permissions were not granted to the correct
quicksight user ARN, as it was using the email value and the users
quicksight username used the User Principal Name (UPN).