Bugfix/revoke permissions (#289)
* Wrap tableaccess delete in a transaction

Changes are to ensure that if deleting the table access fails, the users
original permissions are restored, and the database object remains.
Otherwise the state of the DB and LF could get out of sync.

* Default to local settings when using manage.py

* Allow revoking to silently fail if the permission does not exist

This change allows the revoke_permission method to silently fail if the
permission does not exist. This is useful when revoking permissions that
may have already been revoked, and avoids getting in a state where the
user cannot be granted permissions again.

* Update user usernames

When creating users, use the UPN, preferred username or email address.
This will only occur on initial login. For existing users, a data
migration has been addeded to update the usernames to their email
address, with the domain lowercase.
This resolves a bug where permissions were not granted to the correct
quicksight user ARN, as it was using the email value and the users
quicksight username used the User Principal Name (UPN).

* Bump chart version