Only create resources from the Producer POC account
Temorary change for the POC, so that we dont create resources from other
accounts. As it seems some of our other accounts (ap-dev, ap-data-dev)
may have shared resources with all accounts in the organisation. Creating
resources for these in the POC is confusing and unwanted. We will need to
look into the RAM shares in those dev accounts and clean them up later.